win: Use -O3 in optimize_speed
Pass /clang:-O3 after /O2 in the "optimize_speed" config. This increases
the optimization level for targets in this config from -O2 to -O3 on
Windows.

"/O2" means something slightly different for clang-cl than
"-O2" means for clang!

For clang-cl, "/O2" is short for "/Og /Oi /Ot /Oy /Ob2 /GF /Gy".
"/Ot" is "optimize for speed" and means the same as "-O2" for
clang. The other flags turn on things like -ffunction-sections.

Non-clang-cl clang's -O3 isn't exposed in the clang-cl driver.
But we can use the `/clang:` flag to pass it through.

(See TranslateOptArg() in MSVC.cpp and
CLModePassThroughArgList in Driver.cpp in clang/lib/Driver)
Experimenting with `clang-cl /O2 /clang:-O3 -###` also suggests
that this should do the right thing.)

At least at some point in the past, -O3 made a difference on
non-Windows, so maybe it helps on Windows. But maybe now that
we use PGO everywhere, it doesn't help -- and then we can remove
it on all platforms and get rid of the then-pointless
optimize_speed config. Both would be nice outcomes.

Bug: 329238562
Change-Id: Ie1b6cc396cfaa556618a520ce7a1ac0eb8ebbdf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5378209
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1274873}
NOKEYCHECK=True
GitOrigin-RevId: 95892118af2c7d7423e035a1ec2d257532e10067