Add msan_eager_checks GN arg.
MSan eager checks has ABI implications, so allow this to be configured
with a GN arg. Currently, this defaults to disabled in Chrome C++
code, but enabled in Chrome Rust code, which results in false positive
MSan failures. Adding this GN arg makes it easier to ensure that
instrumented libraries are regenerated with the appropriate build flags.

Bug: 1369167
Change-Id: I00bd6996cdc534b45eacaccaf571e60336ecfc41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5008688
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1281373}
NOKEYCHECK=True
GitOrigin-RevId: 2e02ea2999a25d585896d55949d594ed97304697