Next generation control of unsafe-buffers-usage plugin
Allow per-file control of warnings without clobbering the whole
build. Do this in a way that allows us to mark all dirty files in
individual directories incrementally, then work to remove those
markings.

Default to checking all source files except:
* generated files
* system headers

Opt paths out of being checked by putting them in the paths
file with a `-` at the start of the line:
```
-not_checked/
```

Opt paths into being checked (overriding the above) by putting
them in the paths file with a `+` at the start of the line:
```
+not_checked/actually_checked_though/
```

Individual files can be opted in/out of checks without clobbering
the whole build by using pragmas, which can appear anywhere in the
file and will apply to the whole file, but not get applied to other
files that include it or are included into it.

To opt a file out of checks:
```
// TODO: crbug.com/XYZ - Remove unsafe pointer usage in this file.
#pragma allow_unsafe_buffers
```

To opt a file into checks:
```
#pragma check_unsafe_buffers
```

Bug: 41497066, 40284755
Change-Id: Iefc06aaebc77b2a70c087975f77822fb5a8025f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5426599
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288287}
NOKEYCHECK=True
GitOrigin-RevId: e7db1e3f3da734dc1baebac6b45b924136106d63