Use raw strings for regexp code in resource_utils.py
Starting with Python 3.12, use of unescaped backslash in strings is
reported as a SyntaxWarning, and may become an error at a later time.
Backslashes are commonly found in regexp code.

Using raw strings fixes the syntax problem and it's recommended "for
all but the simplest expressions". Newer code is already doing that.

Bug: chromium:40283283
Change-Id: Ifacf6491d277fc67420d869b7640d8210b0cab40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458533
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288690}
NOKEYCHECK=True
GitOrigin-RevId: 270151221dd9c2dea916bfb439d30e17f4049eaa