[iOS] Bumping the minimum deployment target to iOS 16
In this CL, the minimum version a user needs to be running to use the
application was raised to iOS 16. In addition, one modification to the
screen_ios.mm file was made in order to bring it into compliance with
the new minimum deployment target.

Bug: 40074070
Change-Id: If2e9e48e6f3fc367bb86769c473b13df6c5fe12c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444804
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1293344}
NOKEYCHECK=True
GitOrigin-RevId: 3c96fb7ba3070e8d5fe0e9d8fb4574b98ad5b523