Limit -O2 linker flags to the official build
This speeds up linking of a symbol_level=2 chrome on Linux from ~35-36 seconds to ~18 seconds. This is especially important for incremental
builds where linking dominates.

There's a smaller but significant speedup for the non-symbol_level=2
release build.

Speedometer 3 scores are not observed to move significantly here so
we're expecting most of the -O2 gains to come from the compile steps.

Bug: 339061152
Change-Id: Ief496d947b11445f63fcaa57a8d95aad3396cf62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494203
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297645}
NOKEYCHECK=True
GitOrigin-RevId: c2c0121633fc41d81b69cbb68d1a0ada279b79db