Guard the unsafe-buffers nocompile test with UNSAFE_BUFFERS_BUILD
This define is present only when the plugin is in use. Previously
it was using a buildflags header, but that header was not being
set with the same value, so it would claim the plugin was in use
when it was not. Remove the buildflags header now, as
UNSAFE_BUFFERS_BUILD has replaced it.

R=hans@chromium.org

Bug: 41497066
Change-Id: Ibca1f98ef5ea82a31965e84d05cb8c4747e316fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5527381
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300054}
NOKEYCHECK=True
GitOrigin-RevId: dc1a4a4798d443e98a8f256a1a76d029e6ab6977