[ios] Build chromium code as a framework
Instead of having to include the entire chrome library 4 times in
the output we compile it once as a framework and then jump into
it from various entry points. This matches the same setup we
use on other platforms with the chrome.dll/so.

Teach ios_app_bundle about transparent application bundles which
are used for these BrowserEngineKit extensions which pack all the
frameworks and resources into the parent bundle.

Change-Id: I09a42f9d1bd7a14ab7156a65ad04fded10f6c8ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499634
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301285}
NOKEYCHECK=True
GitOrigin-RevId: 8cda9b45932fa533df5bded387f11d05379ca4bb