Allow a cargo crate to provide a native lib file for Windows
We add `native_libs` as a variable in the cargo_crate template. When
present, any libs pointed to by it will be depended on by the crate.
While depending on them, they will also be copied to the output
directory during build in order to make them visible in the library
search path so that a #[link] directive is able to find them.

We add a gnrt_config.toml option `native_libs_roots` which acts
much like the `extra_input_roots` but for lib files, and is a relative
path to lib files or a subtree of lib files. Any file or directory
listed is walked, and .lib files are collected by gnrt. Then the
paths are written in GN to the `native_libs` GN variable.

This will be tested against the windows-sys crate when it is added.

R=lukasza@chromium.org

Bug: 342194487
Change-Id: I9ceb95cc0ae485cacce3c23a826bfa2b3eae696e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5559449
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1305090}
NOKEYCHECK=True
GitOrigin-RevId: 483a04befc1b7b754d78a22472cbd87e704eb760