android: Pick missing direct from transitive deps
Rather than iterating over all the potential GN deps that could be added
to fulfill the need of a missing class, in the bytecode_processor.py
step we have an added benefit of knowing that javac passed for this
target, which means that the missing class is certain to be found in one
of the transitive deps.

Thus choose a potential direct dep from the list of potentials in the
target's transitive deps rather than the set of all targets that can
provide the missing class.

Also rename variables and dictionaries to make it clear what mapping
they contain and for easier maintainability.

Bug: 341155381
Change-Id: I52e72a0c561b6c5c6b9e21b2b08face4ef147cb1
Fixed: 341155381
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5564328
Commit-Queue: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1306955}
NOKEYCHECK=True
GitOrigin-RevId: 71a0b381f73fda6a1ff6cc781d621be8bb5020b6