Align ubsan_flags and ubsan_security_flags on -fsanitize=shift
ubsan_flags is almost a superset of ubsan_security_flags. The only
difference is ubsan_security_flags uses shift while ubsan_flags uses
shift-exponent, and ubsan_security_flags includes function.

shift vs shift-exponent is easy. shift is shift-exponent + shift-base,
and now that we're on C++20, shift-base never fires. Starting C++20, I
believe `a << b` and `a >> b` is defined for all values of `a`, even
when `a` is signed. (It is not defined for all values of `b`, but that's
covered by shift-exponent.)

Bug: 40942951
Change-Id: I6a1d74fdef5729daf72606a7b12084239bdacd76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588384
Reviewed-by: Titouan Rigoudy <titouan@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309317}
NOKEYCHECK=True
GitOrigin-RevId: 5600def85cbd403c5b99b490aaee0cac4f61ab05