[exception-db] Add metric proto and its implementations
The principle is quite simple and follows common metric systems, but
with several changes.

1) Allow client side aggregation to reduce the load. Recording all raw
data points is still supported, but discouraged due to the in memory
and on-the-wire size.
2) Only support float as the values. (Data points have timestamps
attached.)

Currently Average, Count and DataPoints have been implemented. The
former two calculate the aggregated values; the later one records the
raw data with timestamps.

Everything goes through the Metric class, it tracks all the metric
implementations and dumps them out in its dump() function.
Eventually there will be a dump_to_json() function to write the
TestScriptMetrics protobuf instance into a json file.

Ideally the implementations should belong to a different package, but
likely the package support in chromium is very weak and hard to use
out of the immediate parent folders. So I ended up putting everything
in the same folder for sake of the simplicity.

Bug: 343242386
Change-Id: If0dd44d3c7f03d02af25050a0ea696324f8acce8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5590445
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1310046}
NOKEYCHECK=True
GitOrigin-RevId: 555fa0a5064d633233dd168ee75348be239faa9e