[exception-db] Implement measures.dump
The dump writes the data into a test_script_metrics.jsonpb file in the
argument specified path.

It's worth considering if locking the singleton Metric instance in the
measures module and the registered Measure instances would be
meaningful to block further register or dump. But for now, I'd keep it
as-is for the sake of simplicity.

Bug: 343242386
Change-Id: I2f22d280b1e72137a693704cda0ee09fd4d72ce7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5631004
Commit-Queue: Zijie He <zijiehe@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1315895}
NOKEYCHECK=True
GitOrigin-RevId: 69325684ad90bce3fc461a8b29c0af59c2ceb0a0