[exception-db] Implement TimeConsumption
This is the very simple, one-of TimeConsumption implementation. It
records only once as an AbstractContextManager. Other multiple /
aggregated TimeConsumption implementations would be added later when
necessary.

The use of this class is very simple, as shown in the test:

# ----
with measures.time_consumption('name'):
  do_something()
# ----

Bug: 343242386
Change-Id: I51e7003ab10fbfcf0cd273fce4394765c8fe74fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5637733
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1316622}
NOKEYCHECK=True
GitOrigin-RevId: b312ae824ad30c43ad3bbc73602c270634dbcfd1