Promote most UBSan flags to the ASan configuration on non-Windows
This ratchets these in so we don't regress them. The alignment and null
checks are not yet clear, so keep them on the FYI configuration for now.
To avoid needing to introduce another set of builders, enable them on
the ASan builders.

This does not enable on Windows yet because the bots are not yet
clear. First, -fsanitize=object-size will need to be suppressed due
to https://github.com/llvm/llvm-project/issues/31914. Beyond that,
there are some issues in
https://ci.chromium.org/ui/p/chromium/builders/try/win-asan/11240/overview
that I have not been able to diagnose yet.

Bug: 40248746
Change-Id: I970266c9ea3eb6b8c911f1bec5e29fc6e0198d4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5646784
Reviewed-by: Titouan Rigoudy <titouan@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319068}
NOKEYCHECK=True
GitOrigin-RevId: 36c0df1671dccd3a8f76248d521c918e6c873afc